Dockerise (almost) the microservice
Networking, especially communication between the pika client and RabbitMQ container, needs to be resolved.

In addition, there seems to be a bug in the vectorisation service. This is NOT a fault of celery; it has been tested to make sure of that.

Classification logic still needs to be added, but it'd be wise to solve the problem with the vectorisation beforehand.